+++ ff-core/lib/FF/Options.hs
@@ -110,6 +110,7 @@ data Search = Search
     , inTasks    :: Bool
     , inWikis    :: Bool
     , inContacts :: Bool
+    , inArchived :: Bool
     , limit      :: Maybe Limit
     }
 
@@ -219,11 +220,13 @@ parseOptions h =
         <*> searchT
         <*> searchW
         <*> searchC
+        <*> searchA
         <*> optional limitOption
     searchT = switch $ long "tasks" <> short 't' <> help "Search among tasks"
     searchW = switch $ long "wiki" <> short 'w' <> help "Search among wiki"
     searchC =
         switch $ long "contacts" <> short 'c' <> help "Search among contacts"
+    searchA = switch $ long "archived" <> short 'a' <> help "Search among archived"
     noteid = argument readDocId $
         metavar "ID" <> help "note id" <> completer completeNoteIds
     noteTextArgument = strArgument $ metavar "TEXT" <> help "note text"